separate file for initBtcRelayTokens
